requirement.txt
opencv2 
matplotlib
numpy
pyscreenshot
autopy
pillow
imagemagick or scrot (image backend)
xdotool
tk 

unix-runescape-client


